+++ src/entropic_sampling/entropic.rs
@@ -295,7 +295,7 @@ where Hist: Histogram + HistogramVal<T>,
     {
         while condition(self) {
             self.entropic_step(&mut energy_fn);
-            print_fn(&self);
+            print_fn(self);
         }
     }
 
@@ -326,7 +326,7 @@ where Hist: Histogram + HistogramVal<T>,
     {
         while condition(self) {
             self.entropic_step_acc(&mut energy_fn);
-            print_fn(&self);
+            print_fn(self);
         }
     }